cooking: Cooking with R

Chopping, peeling, frying, and cooking various ingredients, and combining them to a delicious ragout. Also includes buying them from a local supermarket.

buy(what, from = supermarket())

supermarket()

peel(what, ...)

chop(what, ...)

fry(what, ...)

cook(what, ...)

combine(...)

Arguments

what
Main ingredient. If it is the name of an existing file, the first line of the file is read

from
Where to buy the food, default: supermarket()

...
Other ingredients

Value

An object of class food that prints its provenance in a nice fashion.

See also

Useful links:

https://github.com/krlmlr/cooking

Report bugs at https://github.com/krlmlr/cooking/issues

Examples

peel(buy(I("carrots")))

#> peeled carrots, made of
#>   raw carrots, bought at
#>     supermarket

chopped_meat <- chop("raw_meat.csv")
chopped_meat

#> chopped meat, made of
#>   raw meat

fry(chopped_meat, with = I("vegetables"))

#> fried meat, made of
#>   chopped meat, made of
#>     raw meat
#>   vegetables